fix(#624): render ranking UI in the LIVE status renderer + migrate E2E
The ranking UI was built into tabs/status-tab.js (renderStatusTab), but the
unified app renders #t-status via suite/status.js (renderSuiteStatusTab) — so
the feature never appeared in production (player.html redirects to the unified app).

- Extract the ranking UI (ballot + ST aggregate + save + cycle/ballot fetches)
  into a shared module public/js/tabs/status-ranking.js (appendRankingSection).
- Wire it into suite/status.js (the live renderer) AND refactor tabs/status-tab.js
  to use it (de-dup; the two status renderers now share the ranking UI).
- ST aggregate is global (no active char required); player ballot needs the voter.
- Migrate tests/feature-624-...spec.js onto bootApp + goToTab('status') + the
  unified DOM; fix ordeals:{}->[]. E2E 4/4 green on the unified app.

Regression-checked: player.spec.js (13) + fix-50 sheet-status pass. (fix-477
vitae-tally failures are a pre-existing dev issue, unrelated — zero refs to status.)
