+++ public/js/tabs/status-tab.js
@@ -12,7 +12,8 @@
  * City status dots displayed out of 10 (status.city is the combined total).
  */
 
-import { apiGet, apiPut } from '../data/api.js';
+import { apiGet } from '../data/api.js';
+import { appendRankingSection } from './status-ranking.js';
 import { esc, displayName, sortName, clanIcon, covIcon, redactPlayer, discordAvatarUrl, isRedactMode } from '../data/helpers.js';
 import { calcCityStatus } from '../data/accessors.js';
 import { CITY_STATUS_APPELLATIONS } from '../data/constants.js';
@@ -218,104 +219,6 @@ function renderStatusSection(heading, headingIcon, rows, activeId, placeholder) 
   return h;
 }
 
-// ── Clan/Covenant ranking ballot (#624) ──────────────────────────────────────
-// Player picks top-3 of their clan and covenant per cycle; ST-only aggregate.
-const LIVE_CYCLE_STATUSES = ['active', 'game', 'prep'];
-
-function memberOptions(members, selectedId, activeId) {
-  let h = `<option value="">— none —</option>`;
-  for (const c of members) {
-    if (String(c._id) === activeId) continue; // never offer yourself
-    const id = String(c._id);
-    h += `<option value="${esc(id)}"${id === selectedId ? ' selected' : ''}>${esc(displayName(c))}</option>`;
-  }
-  return h;
-}
-
-function renderRankingSlots(rank, label, members, ballotRanking, activeId) {
-  const ords = { 1: '1st', 2: '2nd', 3: '3rd' };
-  let h = `<div class="status-ranking-col"><div class="status-ranking-col-head">${label}</div>`;
-  for (const slot of ['1', '2', '3']) {
-    const sel = ballotRanking && ballotRanking[slot] ? String(ballotRanking[slot]) : '';
-    h += `<label class="status-ranking-slot"><span class="status-ranking-ord">${ords[slot]}</span>`;
-    h += `<select class="status-ranking-sel" data-rank="${rank}" data-slot="${slot}">${memberOptions(members, sel, activeId)}</select></label>`;
-  }
-  return h + `</div>`;
-}
-
-function renderRankingBallot(me, clanMembers, covMembers, ballot, activeId, hasCycle) {
-  let h = `<div class="status-ranking-section">`;
-  h += `<div class="status-section-head"><span class="status-section-title">Clan &amp; Covenant Ranking</span>`;
-  h += `<span class="status-section-caps">your top 3 this cycle — seen only by Storytellers</span></div>`;
-  if (!hasCycle) {
-    h += `<p class="placeholder-msg status-empty">No active downtime cycle to rank in.</p></div>`;
-    return h;
-  }
-  const b = ballot || {};
-  h += `<div class="status-ranking-grid">`;
-  h += renderRankingSlots('clan', `Clan${me.clan ? ' (' + esc(me.clan) + ')' : ''}`, clanMembers, b.clan_ranking, activeId);
-  h += renderRankingSlots('covenant', `Covenant${me.covenant ? ' (' + esc(me.covenant) + ')' : ''}`, covMembers, b.covenant_ranking, activeId);
-  h += `</div>`;
-  h += `<div class="status-ranking-actions"><button class="btn status-ranking-save" type="button">Save Ranking</button><span class="status-ranking-msg" aria-live="polite"></span></div>`;
-  return h + `</div>`;
-}
-
-function renderRankingAggregate(chars, agg) {
-  const byId = new Map(chars.map(c => [String(c._id), c]));
-  const list = (points) => {
-    const rows = Object.entries(points || {})
-      .map(([id, pts]) => ({ c: byId.get(id), pts }))
-      .filter(r => r.c && r.pts > 0)
-      .sort((a, b) => b.pts - a.pts || sortName(a.c).localeCompare(sortName(b.c)));
-    if (!rows.length) return `<p class="placeholder-msg status-empty">No ballots cast yet.</p>`;
-    let h = `<div class="status-ranking-agg-list">`;
-    for (const r of rows) {
-      h += `<div class="status-ranking-agg-row"><span class="status-ranking-agg-name">${esc(displayName(r.c))}</span><span class="status-ranking-agg-pts">${r.pts}</span></div>`;
-    }
-    return h + `</div>`;
-  };
-  let h = `<div class="status-ranking-section">`;
-  h += `<div class="status-section-head"><span class="status-section-title">Ranking Points — this cycle</span>`;
-  h += `<span class="status-section-caps">ST only · 1st=3 2nd=2 3rd=1</span></div>`;
-  h += `<div class="status-ranking-agg-grid">`;
-  h += `<div class="status-ranking-agg-col"><div class="status-ranking-col-head">Clan points</div>${list(agg.clan_points)}</div>`;
-  h += `<div class="status-ranking-agg-col"><div class="status-ranking-col-head">Covenant points</div>${list(agg.covenant_points)}</div>`;
-  return h + `</div></div>`;
-}
-
-function wireRankingSave(el, voterId, cycleId) {
-  const btn = el.querySelector('.status-ranking-save');
-  if (!btn) return;
-  btn.addEventListener('click', async () => {
-    const msg = el.querySelector('.status-ranking-msg');
-    const collect = (rank) => {
-      const picks = {};
-      el.querySelectorAll(`.status-ranking-sel[data-rank="${rank}"]`).forEach(s => { if (s.value) picks[s.dataset.slot] = s.value; });
-      return picks;
-    };
-    const clan = collect('clan');
-    const cov  = collect('covenant');
-    const hasDup = (o) => { const v = Object.values(o); return v.length !== new Set(v).size; };
-    if (hasDup(clan) || hasDup(cov)) { if (msg) msg.textContent = 'Each slot must be a different character.'; return; }
-    if (msg) msg.textContent = 'Saving…';
-    try {
-      await apiPut('/api/ranking_ballots', { cycle_id: cycleId, voter_character_id: voterId, clan_ranking: clan, covenant_ranking: cov });
-      if (msg) msg.textContent = 'Saved.';
-    } catch (err) {
-      if (msg) msg.textContent = 'Save failed: ' + (err.message || 'error');
-    }
-  });
-}
-
-async function resolveActiveCycleId() {
-  try {
-    const cycles = await apiGet('/api/downtime_cycles');
-    const sorted = [...(cycles || [])].sort((a, b) => String(b._id).localeCompare(String(a._id)));
-    const live = sorted.find(c => LIVE_CYCLE_STATUSES.includes(c.status));
-    return live ? String(live._id) : null;
-  } catch { return null; }
-}
-
 // ── Main render ──────────────────────────────────────────────────────────────
 export async function renderStatusTab(el, activeChar, isST = false) {
   if (!el) return;
@@ -335,7 +238,6 @@ export async function renderStatusTab(el, activeChar, isST = false) {
   }
 
   const activeId = String(activeChar._id);
-  const cycleId = await resolveActiveCycleId();
   let h = renderCitySection(chars, activeId);
 
   if (isST) {
@@ -368,13 +270,6 @@ export async function renderStatusTab(el, activeChar, isST = false) {
       h += renderStatusSection(cov, covIcon(cov, 18), rows, activeId, '');
     }
     h += `</div>`;
-
-    // #624: ST-only ranking-points aggregate for the active cycle
-    if (cycleId) {
-      let agg = { clan_points: {}, covenant_points: {} };
-      try { agg = await apiGet(`/api/ranking_ballots/aggregate?cycle_id=${encodeURIComponent(cycleId)}`); } catch { /* ignore */ }
-      h += renderRankingAggregate(chars, agg);
-    }
   } else {
     // Player view — data shaping delegated to data/status-data.js so this
     // file and suite/status.js stay byte-identical on the filter logic.
@@ -411,17 +306,8 @@ export async function renderStatusTab(el, activeChar, isST = false) {
       }
     }
     h += `</div>`;
-
-    // #624: player's clan/covenant ranking ballot for the active cycle
-    const clanMembers = (me?.clan ? clanRowsFor(chars, me.clan, sortName) : []).map(r => r.c).filter(c => String(c._id) !== activeId);
-    const covMembers  = (me?.covenant ? covenantRowsFor(chars, me.covenant, sortName) : []).map(r => r.c).filter(c => String(c._id) !== activeId);
-    let ballot = null;
-    if (cycleId) {
-      try { ballot = await apiGet(`/api/ranking_ballots/mine?cycle_id=${encodeURIComponent(cycleId)}&voter=${encodeURIComponent(activeId)}`); } catch { /* ignore */ }
-    }
-    h += renderRankingBallot(me || activeChar, clanMembers, covMembers, ballot, activeId, !!cycleId);
   }
 
   el.innerHTML = h;
-  if (!isST && cycleId) wireRankingSave(el, activeId, cycleId);
+  await appendRankingSection(el, { chars, activeChar, isST });
 }
